third_party/blink/web_tests/virtual/webnn-service-on-cpu/external/wpt/webnn/idlharness.https.any.worker-expected.txt

This is a testharness.js-based test.
Found 2 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] MLGraphBuilder interface: operation clamp(MLOperand, optional MLClampOptions)
  assert_equals: property has wrong .length expected 0 but got 1
[FAIL] MLGraphBuilder interface: operation clamp(optional MLClampOptions)
  assert_equals: property has wrong .length expected 0 but got 1
Harness: the test ran to completion.
